@kadewu 2024-10-22
> The code looks ok. Could you add any test results of time execution?

Updated.